requirements.txt
Django~=5.1.7
psycopg2
djangorestframework~=3.15.2
faker
celery~=5.4.0
redis~=5.2.1
qrcode
Pillow